Minor changes: + Fix world border issue when transition size change. + Fix command tab completion issue. + Change base structure of PathFinderEntity classes and border wrappers.

+++ api/src/main/java/com/hakan/core/scheduler/Scheduler.java
@@ -105,7 +105,7 @@ public final class Scheduler {
      * @return This class.
      */
     @Nonnull
-    public Scheduler after(int after, @Nonnull TimeUnit timeUnit) {
+    public Scheduler after(long after, @Nonnull TimeUnit timeUnit) {
         this.after = timeUnit.toMillis(after) / 50;
         return this;
     }
@@ -118,7 +118,7 @@ public final class Scheduler {
      * @return This class.
      */
     @Nonnull
-    public Scheduler every(int every, @Nonnull TimeUnit timeUnit) {
+    public Scheduler every(long every, @Nonnull TimeUnit timeUnit) {
         this.every = timeUnit.toMillis(every) / 50;
         return this;
     }
